Added Timestamp module integration to the Contracts module.
