+++ modules/contracts/Cargo.toml
@@ -11,8 +11,10 @@ std = [
 	"parity-codec/std",
 	"runtime_std/std",
 	"primitives/std",
+	"runtime_primitives/std",
 	"support/std",
 	"system/std",
+	"timestamp/std",
 	"oracle/std",
 	"structures/std",
 ]
@@ -40,6 +42,12 @@ git = "https://github.com/paritytech/substrate.git"
 branch = "v1.0"
 default-features = false
 
+[dependencies.runtime_primitives]
+package = "sr-primitives"
+git = "https://github.com/paritytech/substrate.git"
+branch = "v1.0"
+default-features = false
+
 # SRML crates
 [dependencies.support]
 package = "srml-support"
@@ -49,6 +57,12 @@ default-features = false
 
 [dependencies.system]
 package = "srml-system"
+git = "https://github.com/paritytech/substrate.git"
+branch = "v1.0"
+default-features = false
+
+[dependencies.timestamp]
+package = "srml-timestamp"
 git = "https://github.com/paritytech/substrate.git"
 branch = "v1.0"
 default-features = false
@@ -70,7 +84,7 @@ package = "sr-io"
 git = "https://github.com/paritytech/substrate.git"
 branch = "v1.0"
 
-[dev-dependencies.runtime_primitives]
-package = "sr-primitives"
-git = "https://github.com/paritytech/substrate.git"
-branch = "v1.0"
+#[dev-dependencies.runtime_primitives]
+#package = "sr-primitives"
+#git = "https://github.com/paritytech/substrate.git"
+#branch = "v1.0"
