+++ modules/contracts/src/deploy.rs
@@ -11,8 +11,8 @@ impl<T: Trait> Module<T> {
             return Err("Contract ID already exists");
         }
 
-        // TODO: Get current time.
-        let t0 = Time::from_values(1969, 07, 20, 20, 17, 00);
+        // Get current time.
+        let t0 = Time::from_unix(<timestamp::Module<T>>::get().as_());
 
         // Calculating the initial contract state.
         let state;
