+++ modules/contracts/src/init.rs
@@ -3,8 +3,8 @@ use super::*;
 // This function basically performs auto-run for this module.
 impl<T: Trait> Module<T> {
     pub fn init() -> Result {
-        // TODO: Get current time.
-        let t = Time::from_values(1969, 07, 20, 20, 17, 00);
+        // Get current time.
+        let t = Time::from_unix(<timestamp::Module<T>>::get().as_());
 
         // Calculating the initial contract state.
         Self::scheduler_run(t)?;
